+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/remote/client/GRPCRemoteClient.java
@@ -110,7 +110,7 @@ public class GRPCRemoteClient implements RemoteClient {
      *
      * @return a channel when the state to be ready
      */
-    ManagedChannel getChannel() {
+    public ManagedChannel getChannel() {
         return getClient().getChannel();
     }