+++ oap-server/server-query-plugin/status-query-plugin/src/main/java/org/apache/skywalking/oap/query/debug/StatusQueryProvider.java
@@ -56,7 +56,7 @@ public class StatusQueryProvider extends ModuleProvider {
     }
 
     public void prepare() throws ServiceNotProvidedException {
-
+        this.registerServiceImplementation(AlarmStatusQueryService.class, new AlarmStatusQueryService(getManager()));
     }
 
     public void start() throws ServiceNotProvidedException {